UI: Isolate landmark highlight code
